requirements.txt
numpy==1.19.5
matplotlib==3.3.3
einops==0.3.0
librosa==0.8.0
scipy==1.5.4
tqdm
transformer
glob2
tensorboard